# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Documentation",
        "contact": {
            "name": "Blog Ananasik",
            "email": "nurassyl.balgabay@yandex.ru"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign up",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Password comfirmation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Returns data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/posts/{id}/comments": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get comments for post",
                "operationId": "posts.comments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns data"
                    }
                }
            }
        },
        "/api/comments": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "Create comment",
                "operationId": "comments.create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "comment": {
                                        "description": "Content of the comment",
                                        "type": "string"
                                    },
                                    "post_id": {
                                        "description": "post id",
                                        "type": "integer"
                                    },
                                    "parent_id": {
                                        "description": "id of the parent comment(send if answer)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "List of posts",
                "operationId": "posts.list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page num",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns data"
                    }
                }
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Create post",
                "operationId": "posts.create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title",
                                        "type": "string"
                                    },
                                    "short_description": {
                                        "description": "short_description",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "content",
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "description": "category_id",
                                        "type": "int"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/posts/{id}/react": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "like/dislike post",
                "operationId": "posts.react",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reaction",
                        "in": "query",
                        "description": "0(dislike) or 1(like) or null to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}